fix(srt): fix a package error when PES_packet_length more than 65535

+++ core/src/main/java/io/github/thibaultbee/streampack/internal/muxers/ts/packets/Pes.kt
@@ -43,7 +43,7 @@ class Pes(
 
         val header = PesHeader(
             streamId = fromMimeType(stream.config.mimeType).value,
-            payloadLength = frame.buffer.remaining().toShort(),
+            payloadLength = frame.buffer.remaining(),
             pts = frame.pts,
             dts = frame.dts
         )
